// source/common/cloudbase.d.ts
declare namespace cloudbase {
    interface ICloudbaseAppSecret {
        appAccessKeyId: string;
        appAccessKey: string;
    }

    interface ICloudbaseConfig {
        env: string;
        region?: string;
        timeout?: number;
        persistence?: cloudbase.auth.Persistence;
        appSecret?: ICloudbaseAppSecret;
        appSign?: string;
        debug?: boolean;
    }

    interface ICloudbaseExtension {
        name: string;
        invoke(opts: any, app: cloudbase.app.App): Promise<any>;
    }

    interface Listeners {
        [key: string]: Function[];
    }

    interface ICloudbaseEvent {
        name: string;
        target: any;
        data: any;
    }

    interface ICloudbaseEventEmitter {
        on(name: string, listener: Function): this;
        off(name: string, listener: Function): this;
        fire(event: string | ICloudbaseEvent, data?: any): this;
    }

    interface ICloudbaseComponent {
        name: string;
        entity: any;
        namespace?: string;
        injectEvents?: {
            bus: ICloudbaseEventEmitter,
            events: string[];
        };
        IIFE?: boolean
    }

    interface ICloudbaseHook {
        entity: any;
        target: string;
    }
    /**
     * 初始化Cloudbase
     *
     * @example
     * ```javascript
     * const app = cloudbase.init({
     *   env: 'your-envid',
     *   timeout: 15000,
     *   appSign: 'your-appSign',
     *   appSecret: {
     *     appAccessKeyId: 'your-appAccessKeyId'
     *     appAccessKey: 'your-appAccessKey'
     *   }
     * });
     * ```
     *
     * @param config 初始化配置
     * @param config.env 环境ID
     * @param config.timeout 【可选】网络请求超时上限，单位`ms`，默认值`15000`
     * @param config.appSign 【可选】安全应用标识，在非浏览器环境下必须配置，文档{@link https://docs.cloudbase.net/api-reference/webv2/adapter.html#jie-ru-liu-cheng}
     * @param config.appSecret 【可选】安全应用凭证，在非浏览器环境下必须配置，文档{@link https://docs.cloudbase.net/api-reference/webv2/adapter.html#jie-ru-liu-cheng}
     *
     * @return {!cloudbase.app.App} 初始化成功的Cloudbase实例
     */
    function init(config: ICloudbaseConfig): cloudbase.app.App;
    /**
     * 使用适配器，使用方式参考 {@link https://docs.cloudbase.net/api-reference/webv2/adapter.html#di-1-bu-an-zhuang-bing-yin-ru-gua-pei-qi}
     *
     * @example
     * ```javascript
     * cloudbase.useAdapters(adapter); // 使用单个适配器
     * cloudbase.useAdapters([         // 使用多个适配器
     *   adapterA,
     *   adapterB
     * ]);
     * ```
     *
     * @param adapters 适配器对象，入参可以为单个适配器对象，也可以是多个适配器对象的数组
     */
    function useAdapters(adapters: any): void;
    /**
     * 注册扩展能力插件，使用方式参考 {@link https://docs.cloudbase.net/extension/abilities/image-examination.html#shi-yong-kuo-zhan}
     *
     * @example
     * ```javascript
     * cloudbase.registerExtension(ext);
     * ```
     *
     * @param ext 扩展能力插件对象
     */
    function registerExtension(ext: ICloudbaseExtension): void;
    /**
     * 【谨慎操作】注册SDK的版本
     *
     * @example
     * ```javascript
     * cloudbase.registerVersion('1.2.1');
     * ```
     *
     * @param version SDK版本
     */
    function registerVersion(version: string): void;
    /**
     * 【谨慎操作】注册SDK的名称
     *
     * @example
     * ```javascript
     * cloudbase.registerSdkName('cloudbase-js-sdk');
     * ```
     *
     * @param name SDK名称
     */
    function registerSdkName(name: string): void;
    /**
     * 【谨慎操作】修改SDK请求的云开发服务地址
     *
     * @example
     * ```javascript
     * cloudbase.registerEndPoint('url','https');
     * ```
     *
     * @param url 服务地址
     * @param protocol 【可选】强制使用某种协议，默认与主站协议一致
     */
    function registerEndPoint(url: string, protocol?: 'http' | 'https'): void;
    /**
     * 【谨慎操作】注册功能模块
     *
     * @example
     * ```javascript
     * cloudbase.registerComponent({});
     * ```
     *
     * @param component 功能模块对象
     */
    function registerComponent(component: ICloudbaseComponent): void;
    /**
     * 【谨慎操作】注册hook
     *
     * @example
     * ```javascript
     * cloudbase.registerHook({});
     * ```
     *
     * @param hook hook对象
     */
    function registerHook(hook: ICloudbaseHook): void;
}
// ...
    type Persistence = 'local' | 'session' | 'none';